tenx_genomics_utils: add unit tests for 'run_cellranger_mkfastq' (& minor updates to existing tests).

+++ auto_process_ngs/test/test_tenx_genomics.py
@@ -6,7 +6,13 @@ import unittest
 import tempfile
 import os
 import shutil
+from bcftbx.mock import MockIlluminaRun
+from auto_process_ngs.mock import MockBcl2fastq2Exe
+from auto_process_ngs.mock import MockCellrangerExe
 from auto_process_ngs.tenx_genomics_utils import *
+
+# Set to False to keep test output dirs
+REMOVE_TEST_OUTPUTS = True
 
 class TestFlowCellId(unittest.TestCase):
     """
@@ -83,7 +89,8 @@ Lane,Sample_ID,Sample_Name,Sample_Plate,Sample_Well,I7_Index_ID,index,Sample_Pro
         self.wd = tempfile.mkdtemp(suffix="TestHasChromiumSCIndices")
     def tearDown(self):
         # Remove temp dir
-        shutil.rmtree(self.wd)
+        if REMOVE_TEST_OUTPUTS:
+            shutil.rmtree(self.wd)
     def _make_sample_sheet(self,content):
         # Make a sample sheet with supplied content in
         # the temporary area
@@ -126,7 +133,8 @@ class TestCellrangerInfo(unittest.TestCase):
         # Restore the PATH env var
         os.environ['PATH'] = self.original_path
         # Remove temp dir
-        shutil.rmtree(self.wd)
+        if REMOVE_TEST_OUTPUTS:
+            shutil.rmtree(self.wd)
     def _make_mock_cellranger_201(self):
         # Make a fake cellranger 2.0.1 executable
         cellranger_201 = os.path.join(self.wd,"cellranger")
@@ -144,12 +152,145 @@ class TestCellrangerInfo(unittest.TestCase):
 
 class TestMetricsSummary(unittest.TestCase):
     """
+    Tests for the 'MetricsSummary' class
     """
     def test_metrics_summary(self):
-        """
+        """MetricsSummary: check estimated number of cells is extracted
         """
         metrics_summary = """Estimated Number of Cells,Mean Reads per Cell,Median Genes per Cell,Number of Reads,Valid Barcodes,Reads Mapped Confidently to Transcriptome,Reads Mapped Confidently to Exonic Regions,Reads Mapped Confidently to Intronic Regions,Reads Mapped Confidently to Intergenic Regions,Reads Mapped Antisense to Gene,Sequencing Saturation,Q30 Bases in Barcode,Q30 Bases in RNA Read,Q30 Bases in Sample Index,Q30 Bases in UMI,Fraction Reads in Cells,Total Genes Detected,Median UMI Counts per Cell
 "2,272","107,875","1,282","245,093,084",98.3%,69.6%,71.9%,6.1%,3.2%,4.4%,51.3%,98.5%,79.2%,93.6%,98.5%,12.0%,"16,437","2,934"
 """
         m = MetricsSummary(metrics_summary)
         self.assertEqual(m.estimated_number_of_cells,2272)
+
+class TestRunCellrangerMkfastq(unittest.TestCase):
+    """
+    Tests for the 'run_cellranger_mkfastq' function
+    """
+    def setUp(self):
+        # Create a temp working dir
+        self.wd = tempfile.mkdtemp(suffix='TestRunCellrangerMkfastq')
+        # Create a temp 'bin' dir
+        self.bin = os.path.join(self.wd,"bin")
+        os.mkdir(self.bin)
+        # Store original location
+        self.pwd = os.getcwd()
+        # Store original PATH
+        self.path = os.environ['PATH']
+        # Move to working dir
+        os.chdir(self.wd)
+        # Placeholders for test objects
+        self.ap = None
+
+    def tearDown(self):
+        # Delete autoprocessor object
+        if self.ap is not None:
+            del(self.ap)
+        # Return to original dir
+        os.chdir(self.pwd)
+        # Restore PATH
+        os.environ['PATH'] = self.path
+        # Remove the temporary test directory
+        if REMOVE_TEST_OUTPUTS:
+            shutil.rmtree(self.wd)
+
+    def test_run_cellranger_mkfastq(self):
+        """run_cellranger_mkfastq: check cellranger is executed
+        """
+        # Create mock source data
+        illumina_run = MockIlluminaRun(
+            "171020_SN7001250_00002_AHGXXXX",
+            "hiseq",
+            top_dir=self.wd)
+        illumina_run.create()
+        # Mock sample sheet with chromium indices
+        sample_sheet_file = os.path.join(self.wd,"samplesheet.csv")
+        with open(sample_sheet_file,'w') as fp:
+            fp.write("""[Header]
+IEMFileVersion,4
+
+[Reads]
+76
+76
+
+[Settings]
+Adapter,AGATCGGAAGAGCACACGTCTGAACTCCAGTCA
+AdapterRead2,AGATCGGAAGAGCGTCGTGTAGGGAAAGAGTGT
+
+[Data]
+Lane,Sample_ID,Sample_Name,Sample_Plate,Sample_Well,I7_Index_ID,index,Sample_Project,Description
+1,smpl1,smpl1,,,A001,SI-GA-A1,10xGenomics,
+2,smpl2,smpl2,,,A005,SI-GA-B1,10xGenomics,
+3,smpl3,smpl3,,,A006,SI-GA-C1,10xGenomics,
+4,smpl4,smpl4,,,A007,SI-GA-D1,10xGenomics,
+""")
+        # Create mock bcl2fastq and cellranger executables
+        MockBcl2fastq2Exe.create(os.path.join(self.bin,"bcl2fastq"))
+        MockCellrangerExe.create(os.path.join(self.bin,"cellranger"))
+        os.environ['PATH'] = "%s:%s" % (self.bin,
+                                        os.environ['PATH'])
+        # Output dir
+        output_dir = "bcl2fastq"
+        self.assertFalse(os.path.exists("HGXXXX"))
+        self.assertFalse(os.path.exists("cellranger_qc_summary.html"))
+        self.assertFalse(os.path.exists(output_dir))
+        # Run 'cellranger mkfastq'
+        exit_code = run_cellranger_mkfastq(sample_sheet_file,
+                                           illumina_run.dirn,
+                                           output_dir)
+        # Check outputs
+        self.assertEqual(exit_code,0)
+        self.assertTrue(os.path.isdir("HGXXXX"))
+        self.assertTrue(os.path.isfile("cellranger_qc_summary.html"))
+        self.assertTrue(os.path.isdir(output_dir))
+
+    def test_run_cellranger_mkfastq_subset_of_lanes(self):
+        """run_cellranger_mkfastq: check cellranger is executed for subset of lanes
+        """
+        # Create mock source data
+        illumina_run = MockIlluminaRun(
+            "171020_SN7001250_00002_AHGXXXX",
+            "hiseq",
+            top_dir=self.wd)
+        illumina_run.create()
+        # Mock sample sheet with chromium indices
+        sample_sheet_file = os.path.join(self.wd,"samplesheet.csv")
+        with open(sample_sheet_file,'w') as fp:
+            fp.write("""[Header]
+IEMFileVersion,4
+
+[Reads]
+76
+76
+
+[Settings]
+Adapter,AGATCGGAAGAGCACACGTCTGAACTCCAGTCA
+AdapterRead2,AGATCGGAAGAGCGTCGTGTAGGGAAAGAGTGT
+
+[Data]
+Lane,Sample_ID,Sample_Name,Sample_Plate,Sample_Well,I7_Index_ID,index,Sample_Project,Description
+1,smpl1,smpl1,,,A001,SI-GA-A1,10xGenomics,
+2,smpl2,smpl2,,,A005,SI-GA-B1,10xGenomics,
+3,smpl3,smpl3,,,A006,SI-GA-C1,10xGenomics,
+4,smpl4,smpl4,,,A007,SI-GA-D1,10xGenomics,
+""")
+        # Create mock bcl2fastq and cellranger executables
+        MockBcl2fastq2Exe.create(os.path.join(self.bin,"bcl2fastq"))
+        MockCellrangerExe.create(os.path.join(self.bin,"cellranger"))
+        os.environ['PATH'] = "%s:%s" % (self.bin,
+                                        os.environ['PATH'])
+        # Output dir
+        output_dir = "bcl2fastq"
+        self.assertFalse(os.path.exists("HGXXXX_34"))
+        self.assertFalse(os.path.exists("cellranger_qc_summary_34.html"))
+        self.assertFalse(os.path.exists(output_dir))
+        # Run 'cellranger mkfastq'
+        exit_code = run_cellranger_mkfastq(sample_sheet_file,
+                                           illumina_run.dirn,
+                                           output_dir,
+                                           lanes="3,4")
+        # Check outputs
+        self.assertEqual(exit_code,0)
+        self.assertTrue(os.path.isdir("HGXXXX_34"))
+        self.assertTrue(os.path.isfile("cellranger_qc_summary_34.html"))
+        self.assertTrue(os.path.isdir(output_dir))
